gitignore test data files
A note on this commit:

I previously made the unwise choice to commit a bunch of large test
image files to the repo. I then realized that I didn't want those
files in the history forevermore, so I went back and reconstructed
the history without them using edited patches. Lesson learned. You
are now living on the new timeline.