Parallelize 2D FFT computation with Rayon
Uses the fact that the 2D computation does 1D FFT on matrix rows.
There is still a significant time required in transposing the matrix,
which is done serially and mostly involves memory reads and writes.
However, this parallization yields a ~2x speedup on the test machine
for an ~ 8k x 4k image.